introduce so3_function
